mv building of console to console module
